Rename hrit_goes reader to goes-imager_hrit
