+++ satpy/readers/goes_imager_hrit.py
