+++ satpy/readers/nc_goes.py
@@ -167,8 +167,8 @@ import xarray.ufuncs as xu
 import pyresample.geometry
 from satpy import CHUNK_SIZE
 from satpy.readers.file_handlers import BaseFileHandler
-from satpy.readers.hrit_goes import (SPACECRAFTS, EQUATOR_RADIUS, POLE_RADIUS,
-                                     ALTITUDE)
+from satpy.readers.goes_imager_hrit import (SPACECRAFTS, EQUATOR_RADIUS, POLE_RADIUS,
+                                            ALTITUDE)
 from satpy.readers.utils import bbox, get_geostationary_angle_extent
 
 logger = logging.getLogger(__name__)
